Director/pyproject.toml
[project]
name = "director"
version = "0.1.0"
description = "Интерактивный редактор видео с искусственным интеллектом"
requires-python = ">=3.13"
dependencies = [
    "PyQt6>=6.6.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "reactivex>=4.0.0",
]

[project.scripts]
director = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
